+++ test/host/esp_heap_caps.h
@@ -1,0 +1,4 @@
+#ifndef ESP_HEAP_CAPS_H
+#define ESP_HEAP_CAPS_H
+
+#endif // ESP_HEAP_CAPS_H