+++ test/host/esp_wifi.h
@@ -1,0 +1,4 @@
+#ifndef ESP_WIFI_H
+#define ESP_WIFI_H
+
+#endif // ESP_WIFI_H